cobo_api_callback_server/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cobo-waas2==1.24.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.0
email-validator==2.3.0
fastapi==0.103.2
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pika==1.3.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
